feat(cli): add 'cd' step to next steps for all pathways
Add a clear 'Change to project directory' step as the first item in
the next steps output for all pathways. This helps users know they
need to cd into the project before running other commands.

Also added next steps output to the non-interactive mode which was
previously missing this information.